requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Azure OpenAI and AI dependencies  
openai>=1.3.0
azure-openai>=1.0.0
azure-identity>=1.15.0
azure-core>=1.29.0

# HTTP client and web scraping
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# Data processing and utilities
python-multipart>=0.0.6
python-json-logger>=2.0.0
python-dotenv>=1.0.0

# Development and testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing FastAPI endpoints

# Optional: Enhanced content analysis
textstat>=0.7.3  # For readability analysis
langdetect>=1.0.9  # For language detection
dateutil>=2.8.2  # For date parsing